project/07.SingleAgent/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
rank-bm25>=0.2.2
numpy>=1.26.0
pypdf>=4.2.0
tavily-python>=0.3.3
python-dotenv>=1.0.1
